added code structure changes
