+++ models/index.js
@@ -1,0 +1,5 @@
+const contactsOperations = require('./contacts')
+
+module.exports = {
+  contactsOperations,
+}
